Move more traits from Mods.RA to Mods.Common
